refactor(db): split heroes.ts into focused modules behind a barrel
heroes.ts had grown to 1,261 lines spanning lookups, search, the
home/explore feed, category browse pages, the relationship graph, and
row→CharacterData transforms — so any search or edit had to wade through
the whole file. Split the implementation into cohesive modules under
src/lib/db/heroes/ and turn heroes.ts into a 16-line barrel that
re-exports them:

  types.ts          shared types (Hero, HeroSearchResult, CategorySlug, …)
  core.ts           by-id lookups, search/browse RPC, pure filters, count
  feed.ts           home/explore feed rows (popular, villains, spotlight, …)
  categories.ts     category browse pages, covers, era timeline, labels
  relationships.ts  related/rivalries/most-feared, power-range, publisher counts
  transforms.ts     heroRowToCharacterData, getHeroFamily

All 51 import sites keep importing from '../lib/db/heroes' unchanged —
the barrel preserves the public API. Pure code-move; no behavior change.
typecheck clean; format clean; 363 tests pass; 0 lint errors.
